Activate Triple Ratchet (Mode 3) with ML-KEM-768 post-quantum ratchet
- Enable MODE3_ENABLED for Zerion-to-Zerion contacts
- Add PQ epoch lifecycle: ML-KEM-768 key exchange every 25 messages or 24 hours
- Add chunked PQ transmission (256-byte chunks piggybacked on messages)
- Mix PQ secret into root key on epoch completion
- Add Mode 3 capability negotiation in handshake
- Add mode3Capable flag to Contact and database schema (migrations 57-60)
- Update PCS stream encrypter/decrypter for Mode 3 headers
- Optimize polling intervals (30s min, 5min max)
- Remove redundant POST_QUANTUM_MESSAGING.md doc
- Update documentation for Triple Ratchet completion

+++ bramble-core/src/test/java/org/briarproject/bramble/record/RecordWriterImplTest.java
@@ -40,7 +40,7 @@ public class RecordWriterImplTest extends BrambleTestCase {
 		assertEquals(RECORD_HEADER_BYTES + payloadLength, written.length);
 		assertEquals(protocolVersion, written[0]);
 		assertEquals(recordType, written[1]);
-		assertEquals(payloadLength, ByteUtils.readUint16(written, 2));
+		assertEquals(payloadLength, ByteUtils.readUint32(written, 2));
 		byte[] writtenPayload = new byte[payloadLength];
 		System.arraycopy(written, RECORD_HEADER_BYTES, writtenPayload, 0,
 				payloadLength);
